+++ frontend/package.json
@@ -20,7 +20,7 @@
     "react-scripts": "4.0.3",
     "styled-components": "^5.2.1",
     "styled-reset": "^4.3.4",
-    "typescript": "^4.1.2",
+    "typescript": "^4.2.2",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
