+++ frontend/src/components/TextElements.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import * as colors from '../../style/colors';
+import * as colors from '../style/colors';
 
 export const StyledTitle = styled.h1`
   font-size: 2rem;
@@ -12,3 +12,13 @@ export const StyledTitle = styled.h1`
     margin-bottom: 3rem;
   }
 `;
+
+export function Title(props: any) {
+  return (
+    <StyledTitle className={props.className}>
+      <span>{props.children}</span>
+      <hr />
+    </StyledTitle>
+  );
+}
+
