+++ frontend/src/style/colors.tsx
@@ -1,4 +1,5 @@
 export const primary = '#0190ab';
+export const primaryShadow = 'rgba(1, 143, 171, 0.5)';
 export const primaryHover = '#017b94';
 export const componentBackground = '#f2f3f5';
 export const background = '#f2f3f5';
